# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TokenSymbol": {
				"type": "string"
			},
...
			"ErrorResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"originalErrorMessage": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"MarketNetwork": {
				"properties": {
					"iconUrl": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"depositEnabled": {
						"type": "boolean"
					},
					"withdrawalEnabled": {
						"type": "boolean"
					},
					"withdrawMin": {
						"type": "string"
					},
					"withdrawMax": {
						"type": "string"
					},
					"withdrawFee": {
						"type": "string"
					},
					"network": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"iconUrl",
					"depositEnabled",
					"withdrawalEnabled",
					"network",
					"name"
				],
				"type": "object"
			},
			"TokenWithMarketNetwork": {
				"properties": {
					"contracts": {
						"$ref": "#/components/schemas/Record_string.string_"
					},
					"precision": {
						"type": "number",
						"format": "double"
					},
					"networks": {
						"items": {
							"$ref": "#/components/schemas/MarketNetwork"
						},
						"type": "array"
					},
					"isWithdrawable": {
						"type": "boolean"
					},
					"isDepositable": {
						"type": "boolean"
					},
					"tokenIcon": {
						"type": "string"
					},
					"tokenSymbol": {
						"$ref": "#/components/schemas/TokenSymbol"
					}
				},
				"required": [
					"networks",
					"isWithdrawable",
					"isDepositable",
					"tokenSymbol"
				],
				"type": "object"
			},
...
			"Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58_PortfolioBalanceEntry___": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"TokenBalanceV2": {
				"properties": {
					"refFreeBalance": {
						"type": "string"
					},
					"freeBalance": {
						"type": "string"
					},
					"refTotalBalance": {
						"type": "string"
					},
					"totalBalance": {
						"type": "string"
					},
					"tokenName": {
						"type": "string"
					}
				},
				"required": [
					"refFreeBalance",
					"freeBalance",
					"refTotalBalance",
					"totalBalance"
				],
				"type": "object"
			},
			"GroupedBalancesV2": {
				"properties": {},
				"additionalProperties": {
					"properties": {
						"token": {
							"$ref": "#/components/schemas/TokenWithMarketNetwork"
						},
						"balance": {
							"$ref": "#/components/schemas/TokenBalanceV2"
						}
					},
					"required": [
						"token",
						"balance"
					],
					"type": "object"
				},
				"type": "object"
			},
			"ReturnType_CedeSDK-at-api_91_getWithdrawableBalancesV2_93__": {
				"$ref": "#/components/schemas/GroupedBalancesV2",
				"description": "Obtain the return type of a function type"
			},
			"GetWithdrawableBalancesResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getWithdrawableBalancesV2_93__"
			},
			"WithdrawableBalancesWithTokensV2": {
				"properties": {},
				"additionalProperties": {
					"properties": {
						"token": {
							"$ref": "#/components/schemas/TokenWithMarketNetwork"
						},
						"balance": {
							"$ref": "#/components/schemas/TokenBalanceV2"
						}
					},
					"required": [
						"token",
						"balance"
					],
					"type": "object"
				},
				"type": "object"
			},
			"ReturnType_CedeSDK-at-api_91_getWithdrawableBalancesWithTokensV2_93__": {
				"$ref": "#/components/schemas/WithdrawableBalancesWithTokensV2",
				"description": "Obtain the return type of a function type"
			},
			"GetWithdrawableBalancesWithTokensResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getWithdrawableBalancesWithTokensV2_93__"
			},
			"ReturnType_CedeSDK-at-api_91_getBalancesV2_93__": {
				"$ref": "#/components/schemas/Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58_PortfolioBalanceEntry___",
				"description": "Obtain the return type of a function type"
			},
			"GetBalancesResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getBalancesV2_93__"
			},
			"Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58__balance-PortfolioBalanceEntry--token-TokenWithMarketNetwork____": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ReturnType_CedeSDK-at-api_91_getBalancesWithTokensV2_93__": {
				"$ref": "#/components/schemas/Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58__balance-PortfolioBalanceEntry--token-TokenWithMarketNetwork____",
				"description": "Obtain the return type of a function type"
			},
			"GetBalancesWithTokensResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getBalancesWithTokensV2_93__"
			},
			"Record_string.SubAccountBalancesWithTokensV2_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58__balance-PortfolioBalanceEntry--token-TokenWithMarketNetwork____"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"BalancesWithTokensV2": {
				"$ref": "#/components/schemas/Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58__balance-PortfolioBalanceEntry--token-TokenWithMarketNetwork____"
			},
			"MainSubAccountsBalancesWithTokensResponseV2": {
				"properties": {
					"subAccounts": {
						"$ref": "#/components/schemas/Record_string.SubAccountBalancesWithTokensV2_"
					},
					"main": {
						"$ref": "#/components/schemas/BalancesWithTokensV2"
					}
				},
				"required": [
					"subAccounts",
					"main"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReturnType_CedeSDK-at-api_91_getMainSubAccountsBalancesWithTokensV2_93__": {
				"$ref": "#/components/schemas/MainSubAccountsBalancesWithTokensResponseV2",
				"description": "Obtain the return type of a function type"
			},
			"GetMainSubAccountsBalancesWithTokensResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getMainSubAccountsBalancesWithTokensV2_93__"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@cedelabs/sdk-http-api",
		"version": "1.4.1",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/portfolio/withdrawable-balances": {
			"get": {
				"operationId": "GetWithdrawableBalances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetWithdrawableBalancesResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get withdrawable balances for an exchange.\nRetrieves balances from a wallet used to initiate withdrawals.\n\nIf you have funds on other wallets, you'll first need to transfer these funds to the withdrawal wallet:\n- you can retrieve withdrawal wallets using `/supported` endpoint (`sendWalletTypes` field)\n\nThe response includes balances grouped by wallet type.",
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/portfolio/withdrawable-balances-with-tokens": {
			"get": {
				"operationId": "GetWithdrawableBalancesWithTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetWithdrawableBalancesWithTokensResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get withdrawable balances with token metadata for an exchange.\nRetrieves balances from a wallet used to initiate withdrawals and provides token metadata (e.g. token icon, contract address if available, etc.).\n\nIf you have funds on other wallets, you'll first need to transfer these funds to the withdrawal wallet:\n- you can retrieve withdrawal wallets using `/supported` endpoint (`sendWalletTypes` field)",
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/portfolio/balances": {
			"get": {
				"operationId": "GetBalances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetBalancesResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all balances for an exchange.\nRetrieves all balances across different wallet types.\nUses a hydration system to deliver cached data instantly while fetching latest balances.",
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/portfolio/balances-with-tokens": {
			"get": {
				"operationId": "GetBalancesWithTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetBalancesWithTokensResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all balances with token metadata for an exchange.\nRetrieves all balances across different wallet types and provides token metadata (e.g. token icon, contract address if available, etc.).",
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/portfolio/main-sub-accounts-balances-with-tokens": {
			"get": {
				"operationId": "GetMainSubAccountsBalancesWithTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetMainSubAccountsBalancesWithTokensResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all main and sub accounts balances with token metadata for an exchange.\nRetrieves all balances across different wallet types and provides token metadata (e.g. token icon, contract address if available, etc.).",
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1"
		}
	]
}